Moved GrimRenderer to mmstd_moldyn
